Adds Twitter and Slack strategies to the refactor
